# packages/api-spec/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ProjectBeacon API",
    "version": "0.0.0",
    "description": "Generated fragment: SessionBrief, comments, decisions, errors, and pagination."
  },
  "servers": [
    {
      "url": "/"
    }
  ],
  "paths": {
...
    "/v1/decisions/{id}": {
      "patch": {
        "summary": "Update a decision lifecycle status",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "pattern": "^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-7[0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}$"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PatchDecision"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated decision",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Decision"
                }
              }
            }
          },
          "default": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ErrorResponse": {
        "$schema": "http://json-schema.org/draft-07/schema#",
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string",
                "enum": [
                  "unauthorized",
                  "invalid_request",
                  "forbidden",
                  "not_found",
                  "version_conflict",
                  "task_locked",
                  "finish_work_required",
                  "dependency_cycle",
                  "login_taken",
                  "repo_ambiguous",
                  "unsupported_media",
                  "rate_limited",
                  "code_index_unavailable",
                  "bootstrap_consumed",
                  "integration_unavailable"
                ]
              },
              "message": {
                "type": "string"
              },
              "details": {
                "default": {},
                "type": "object",
                "propertyNames": {
                  "type": "string"
                },
                "additionalProperties": {}
              }
            },
            "required": [
              "code",
              "message",
              "details"
            ],
            "additionalProperties": false
          }
        },
        "required": [
          "error"
        ],
        "additionalProperties": false
      },
...
      "Decision": {
        "$schema": "http://json-schema.org/draft-07/schema#",
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-7[0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12})$"
          },
          "project_id": {
            "type": "string",
            "format": "uuid",
            "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-7[0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12})$"
          },
          "title": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "proposed",
              "accepted",
              "superseded",
              "deprecated"
            ]
          },
          "context": {
            "type": "string"
          },
          "decision": {
            "type": "string"
          },
          "consequences": {
            "type": "string"
          },
          "created_by_type": {
            "type": "string"
          },
          "created_by_id": {
            "type": "string"
          },
          "superseded_by": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid",
                "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-7[0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12})$"
              },
              {
                "type": "null"
              }
            ]
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z|([+-](?:[01]\\d|2[0-3]):[0-5]\\d)))$"
          },
          "related_paths": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "repo_id": {
                  "type": "string",
                  "format": "uuid",
                  "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-7[0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12})$"
                },
                "path": {
                  "type": "string"
                }
              },
              "required": [
                "repo_id",
                "path"
              ],
              "additionalProperties": false
            }
          },
          "related_task_ids": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid",
              "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-7[0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12})$"
            }
          }
        },
        "required": [
          "id",
          "project_id",
          "title",
          "status",
          "context",
          "decision",
          "consequences",
          "created_by_type",
          "created_by_id",
          "superseded_by",
          "created_at",
          "related_paths",
          "related_task_ids"
        ],
        "additionalProperties": false
      },
      "PatchDecision": {
        "$schema": "http://json-schema.org/draft-07/schema#",
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "accepted",
              "superseded",
              "deprecated"
            ]
          },
          "superseded_by": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid",
                "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-7[0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12})$"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "status"
        ],
        "additionalProperties": false
      }
    }
  }
}